requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
deep-translator==1.11.4
detoxify==0.5.2
Django==5.2.4
dotenv==0.9.9
filelock==3.18.0
flatbuffers==25.9.23
fonttools==4.59.0
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.34.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keybert==0.9.0
kiwisolver==1.4.8
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nudenet==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
Pygments==2.19.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.0.0
sentencepiece==0.2.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch>=2.1.0
torchvision>=0.16.0
tqdm==4.67.1
transformers>=4.30.0
tweepy==4.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-ratelimit==4.1.0
stripe==11.3.0
